Move back fireweather_bot.py to where it was
